cli: Refactor git handlers out of init module
